Implement deck management and card creation UI
The old deck management UI was migrated from dedicated data editing pages (e.g., "/deck/[id]/edit") to the new approach with sliding modals that open on the same page, making the process of data modification easier to initiate from any part of the application. In the future, this will make it easier to implement card editing directly in search results, for example. The card creation UI uses this new approach from the beginning.