Implement session validation to protected routes
Previously, the only protection layer for authenticated-only routes was the middleware, and the routes relied on the session being valid. Now, when user data is fetched in protected routes, the code accounts for the case where the session is invalid in ways that cannot be caught by the middleware (for example, the token does have a user ID but it does not correspond to an existing user record). In these cases, the sign-out sequence is now initiated.