Make invalid token sign-out seamless
This replaces the old invalid token sign-out flow - redirecting to a route that does not render anything and instead triggers the standard sign-out Server Action - with the new flow. The new flow instead redirects the user to the sign-in page (much like an attempt to access protected routes without any session token would) and adds a special search parameter flag - "invalidToken". This flag causes the page to trigger the sign-out action and, upon completion, erase the flag from the URL to avoid confusing the user. With this flow, the user sees the sign-in page instead of an empty page while the token is being destroyed.