Remove forced Docker container names
This change allows Docker Compose to automatically generate names based on the combination of the project name and the service name. For example, it allows the production project and test project to use the same service names without causing container name collisions because the actual container names also include the project names.