Improve timezone handling for OAuth-based accounts
Due to issues with obtaining time zone information from OAuth providers, the time zone is now inferred for OAuth accounts based on browser settings in the same way as it is for credentials-based accounts. Since there is no direct transition from the client to the server code that creates the associated user record for new OAuth sign-ins (the OAuth provider's authentication flow occurs between them), the time zone is marked as missing (null) initially and lazily inferred upon the next visit to the homepage, which normally happens immediately since the user is redirected directly to it after a successful sign-in.