Fix OAuth user initial tokens getting invalidated
The issue was occurring because the default value (current timestamp) assigned to the "acceptTokensAfter" field had higher precision than the tokens' "issued at" field, which resulted in the "acceptTokensAfter" timestamp technically being in the future and thus invalidating tokens that should be valid. This was fixed by explicitly truncating the token acceptance threshold timestamp to seconds.